FF: fixes to the refactoring of the experiment package
